Extract primitives into separate package
