+++ apps/web/package.json
@@ -79,6 +79,7 @@
     "@types/prettier": "^2.7.2",
     "@unocss/reset": "^0.50.6",
     "@vrite/backend": "workspace:*",
+    "@vrite/components": "workspace:*",
     "@vrite/sdk": "2023.5.21-beta",
     "@vrite/tiptap-solid": "2023.3.24-2",
     "clsx": "^1.2.1",
