+++ apps/web/src/components/primitives/index.ts
@@ -1,0 +1,2 @@
+export * from "./sortable";
+export * from "@vrite/components";
