+++ packages/components/src/primitives/tooltip.tsx
@@ -1,9 +1,9 @@
+import { Fragment } from "./fragment";
+import { createRef } from "../ref";
 import clsx from "clsx";
 import { Component, createEffect, createSignal, JSX, onMount, Show, on, onCleanup } from "solid-js";
 import { computePosition, hide, autoUpdate } from "@floating-ui/dom";
 import { Dynamic, Portal } from "solid-js/web";
-import { createRef } from "#lib/utils";
-import { Fragment } from "#components/primitives";
 
 type TooltipPosition = {
   x: number;
